requirements.txt
fasttext-wheel
nltk